requirements.txt
# Runtime dependencies
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1

# Google ADK and Gemini
google-adk==1.0.0
google-generativeai==0.8.4

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
